@@ -1,7 +1,7 @@
 package de.dkh.customerapi.entity;
 
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
